claude-pdf-analyzer/requirements.txt
streamlit>=1.50.0
anthropic>=0.84.0
langchain>=0.3.0
langchain-community
langchain-anthropic
langchain-huggingface
langchain-chroma
pypdf>=6.0.0
chromadb>=0.5.0
python-dotenv
pysqlite3-binary